+++ presto-main/src/main/java/com/facebook/presto/execution/buffer/OutputBuffer.java
@@ -13,9 +13,8 @@
  */
 package com.facebook.presto.execution.buffer;
 
-import com.facebook.presto.OutputBuffers;
-import com.facebook.presto.OutputBuffers.OutputBufferId;
 import com.facebook.presto.execution.StateMachine.StateChangeListener;
+import com.facebook.presto.execution.buffer.OutputBuffers.OutputBufferId;
 import com.google.common.util.concurrent.ListenableFuture;
 import io.airlift.units.DataSize;
